Don't require manual execution of description_to_json, run during cmake instead
