+++ python/plugins/grassprovider/CMakeLists.txt
@@ -1,6 +1,16 @@
 file(GLOB PY_FILES *.py)
 file(GLOB OTHER_FILES grass7.txt metadata.txt)
-file(GLOB DESCR_FILES description/algorithms.json)
+
+execute_process(
+  COMMAND ${Python_EXECUTABLE} -m grassprovider.description_to_json ${CMAKE_CURRENT_SOURCE_DIR}/description ${CMAKE_CURRENT_SOURCE_DIR}/description/algorithms.json
+  WORKING_DIRECTORY ${CMAKE_CURRENT_SOURCE_DIR}/..
+  RESULT_VARIABLE result
+  ERROR_VARIABLE error_output
+)
+if(NOT "${result}" STREQUAL "0")
+    message(FATAL_ERROR "Create grass provider algorithm descriptions failed with error: ${error_output}")
+endif()
+set(DESCR_FILES ${CMAKE_CURRENT_SOURCE_DIR}/description/algorithms.json)
 
 add_subdirectory(ext)
 add_subdirectory(tests)
