+++ python/plugins/grassprovider/Grass7Utils.py
@@ -44,188 +44,11 @@ from qgis.core import (
     QgsMessageLog,
     QgsCoordinateReferenceSystem,
     QgsProcessingContext,
-    QgsProcessingParameterDefinition,
-    QgsProcessingParameterVectorLayer,
-    QgsProcessingParameterFeatureSource,
-    QgsProcessingParameterRasterLayer,
-    QgsProcessingParameterMultipleLayers,
-    QgsProcessingParameterVectorDestination,
-    QgsProcessingParameterRasterDestination
 )
 
 from processing.algs.gdal.GdalUtils import GdalUtils
 from processing.core.ProcessingConfig import ProcessingConfig
-from processing.core.parameters import getParameterFromString
 from processing.tools.system import userFolder, isWindows, isMac, mkdir
-
-
-@dataclass
-class ParsedDescription:
-    """
-    Results of parsing a description file
-    """
-
-    GROUP_ID_REGEX = re.compile(r'^[^\s(]+')
-
-    grass_command: Optional[str] = None
-    short_description: Optional[str] = None
-    name: Optional[str] = None
-    display_name: Optional[str] = None
-    group: Optional[str] = None
-    group_id: Optional[str] = None
-
-    ext_path: Optional[str] = None
-
-    has_raster_input: bool = False
-    has_vector_input: bool = False
-
-    has_raster_output: bool = False
-    has_vector_outputs: bool = False
-
-    hardcoded_strings: List[str] = field(default_factory=list)
-    params: List[QgsProcessingParameterDefinition] = field(
-        default_factory=list)
-    param_strings: List[str] = field(default_factory=list)
-
-    def as_dict(self) -> Dict:
-        """
-        Returns a JSON serializable dictionary representing the parsed
-        description
-        """
-        return {
-            'name': self.name,
-            'display_name': self.display_name,
-            'command': self.grass_command,
-            'short_description': self.short_description,
-            'group': self.group,
-            'group_id': self.group_id,
-            'ext_path': self.ext_path,
-            'inputs': {
-                'has_raster': self.has_raster_input,
-                'has_vector': self.has_vector_input
-            },
-            'outputs':
-                {
-                    'has_raster': self.has_raster_output,
-                    'has_vector': self.has_vector_outputs
-            },
-            'hardcoded_strings': self.hardcoded_strings,
-            'parameters': self.param_strings
-        }
-
-    @staticmethod
-    def from_dict(description: Dict) -> 'ParsedDescription':
-        """
-        Parses a dictionary as a description and returns the result
-        """
-        result = ParsedDescription()
-        result.name = description.get('name')
-        result.display_name = description.get('display_name')
-        result.grass_command = description.get('command')
-        result.short_description = QCoreApplication.translate(
-            "GrassAlgorithm",
-            description.get('short_description')
-        )
-        result.group = QCoreApplication.translate("GrassAlgorithm",
-                                                  description.get('group'))
-        result.group_id = description.get('group_id')
-        result.ext_path = description.get('ext_path')
-        result.has_raster_input = description.get('inputs', {}).get(
-            'has_raster', False)
-        result.has_vector_input = description.get('inputs', {}).get(
-            'has_vector', False)
-        result.has_raster_output = description.get('outputs', {}).get(
-            'has_raster', False)
-        result.has_vector_outputs = description.get('outputs', {}).get(
-            'has_vector', False)
-        result.hardcoded_strings = description.get('hardcoded_strings', [])
-        result.param_strings = description.get('parameters', [])
-        for param in result.param_strings:
-            result.params.append(
-                getParameterFromString(param, "GrassAlgorithm"))
-
-        return result
-
-    @staticmethod
-    def parse_description_file(
-            description_file: Path,
-            translate: bool = True) -> 'ParsedDescription':
-        """
-        Parses a description file and returns the result
-        """
-        result = ParsedDescription()
-
-        with description_file.open() as lines:
-            # First line of the file is the Grass algorithm name
-            line = lines.readline().strip('\n').strip()
-            result.grass_command = line
-            # Second line if the algorithm name in Processing
-            line = lines.readline().strip('\n').strip()
-            result.short_description = line
-            if " - " not in line:
-                result.name = result.grass_command
-            else:
-                result.name = line[:line.find(' ')].lower()
-            if translate:
-                result.short_description = QCoreApplication.translate(
-                    "GrassAlgorithm", line)
-            else:
-                result.short_description = line
-
-            result.display_name = result.name
-            # Read the grass group
-            line = lines.readline().strip('\n').strip()
-            if translate:
-                result.group = QCoreApplication.translate("GrassAlgorithm",
-                                                          line)
-            else:
-                result.group = line
-
-            result.group_id = ParsedDescription.GROUP_ID_REGEX.search(
-                line).group(0).lower()
-
-            # Then you have parameters/output definition
-            line = lines.readline().strip('\n').strip()
-            while line != '':
-                try:
-                    line = line.strip('\n').strip()
-                    if line.startswith('Hardcoded'):
-                        result.hardcoded_strings.append(
-                            line[len('Hardcoded|'):])
-                    result.param_strings.append(line)
-                    parameter = getParameterFromString(line, "GrassAlgorithm")
-                    if parameter is not None:
-                        result.params.append(parameter)
-                        if isinstance(parameter, (
-                                QgsProcessingParameterVectorLayer,
-                                QgsProcessingParameterFeatureSource)):
-                            result.has_vector_input = True
-                        elif isinstance(parameter,
-                                        QgsProcessingParameterRasterLayer):
-                            result.has_raster_input = True
-                        elif isinstance(parameter,
-                                        QgsProcessingParameterMultipleLayers):
-                            if parameter.layerType() < 3 or parameter.layerType() == 5:
-                                result.has_vector_input = True
-                            elif parameter.layerType() == 3:
-                                result.has_raster_input = True
-                        elif isinstance(parameter,
-                                        QgsProcessingParameterVectorDestination):
-                            result.has_vector_outputs = True
-                        elif isinstance(parameter,
-                                        QgsProcessingParameterRasterDestination):
-                            result.has_raster_output = True
-                    line = lines.readline().strip('\n').strip()
-                except Exception as e:
-                    QgsMessageLog.logMessage(
-                        QCoreApplication.translate("GrassAlgorithm",
-                                                   'Could not open GRASS GIS 7 algorithm: {0}\n{1}').format(
-                            description_file, line),
-                        QCoreApplication.translate("GrassAlgorithm",
-                                                   'Processing'),
-                        Qgis.Critical)
-                    raise e
-        return result
 
 
 class Grass7Utils:
